rag-pipeline-main/rag-pipeline-main/requirements.txt
pyautogen
pyautogen[gemini]
flaml[automl]
